Grids: support compound keys in AI Column (#34520)
(cherry picked from commit 396a1abf0644b8aab8c6b7b56b707492a89b9319)
