+++ packages/devextreme/js/__internal/grids/grid_core/ai_column/utils.ts
@@ -20,11 +20,11 @@ export const getDataFromRowItems = (items: Item[]): UserData[] => items
 export const reduceDataCachedKeys = (
   data: UserData[],
   cachedData: Record<PropertyKey, string>,
-  keyField: string,
+  getKey: (item: UserData) => PropertyKey,
 ): Record<PropertyKey, unknown> => {
   const newData: Record<PropertyKey, unknown> = { };
   for (const item of data) {
-    const key = item[keyField] as PropertyKey;
+    const key = getKey(item);
     if (!(key in cachedData)) {
       newData[key] = item;
     }
@@ -35,15 +35,17 @@ export const reduceDataCachedKeys = (
 
 export const isKeyMissingInData = (
   data: UserData[],
-  keyField: string | string[],
+  keyField: string | string[] | ((data: UserData) => unknown),
 ): boolean => {
-  if (typeof keyField !== 'string') {
-    // The key field should be a string for AI Column functionality.
-    // Return false to avoid unnecessary errors for compound keys.
-    return false;
+  if (typeof keyField === 'function') {
+    return data.some((item) => !isDefined(keyField(item)));
   }
 
-  return data.some((item) => !(keyField in item));
+  const keyFields = Array.isArray(keyField) ? keyField : [keyField];
+
+  return data.some(
+    (item) => keyFields.some((field) => !(field in item)),
+  );
 };
 
 export const isAIColumnAutoMode = (column: Column): boolean => column.type === 'ai' && (!column.ai?.mode || column.ai.mode === 'auto');
